// src/smite/types/get-mach-player-details.ts
export class GetMatchPlayerDetails {
  Account_Gods_Played: number;
  Account_Level: number;
  GodId: number;
  GodLevel: number;
  GodName: string;
  Mastery_Level: number;
  Match: number;
  Queue: string;
  Rank_Stat: number;
  SkinId: number;
  Tier: number;
  mapGame: string;
  playerCreated: string;
  playerId: string;
  playerName: string;
  playerRegion: string;
  ret_msg: any;
  taskForce: number;
  tierLosses: number;
  tierPoints: number;
  tierWins: number;
}