// src/smite/types/get-player.ts
export class GetPlayer {
  ActivePlayerId: number;
  Avatar_URL: string;
  Created_Datetime: string;
  HoursPlayed: number;
  Id: number;
  Last_Login_Datetime: string;
  Leaves: number;
  Level: number;
  Losses: number;
  MasteryLevel: number;
  MergedPlayers: any;
  MinutesPlayed: number;
  Name: string;
  Personal_Status_Message: string;
  Platform: string;
  Rank_Stat_Conquest: number;
  Rank_Stat_Conquest_Controller: number;
  Rank_Stat_Duel: number;
  Rank_Stat_Duel_Controller: number;
  Rank_Stat_Joust: number;
  Rank_Stat_Joust_Controller: number;
  RankedConquest: {
    Leaves: number;
    Losses: number;
    Name: string;
    Points: number;
    PrevRank: number;
    Rank: number;
    Rank_Stat: number;
    Rank_Stat_Conquest: any;
    Rank_Stat_Joust: any;
    Rank_Variance: number;
    Round: number;
    Season: number;
    Tier: number;
    Trend: number;
    Wins: number;
    player_id: any;
    ret_msg: any;
  };
  RankedConquestController: {
    Leaves: number;
    Losses: number;
    Name: string;
    Points: number;
    PrevRank: number;
    Rank: number;
    Rank_Stat: number;
    Rank_Stat_Conquest: any;
    Rank_Stat_Joust: any;
    Rank_Variance: number;
    Round: number;
    Season: number;
    Tier: number;
    Trend: number;
    Wins: number;
    player_id: any;
    ret_msg: any;
  };
  RankedDuel: {
    Leaves: number;
    Losses: number;
    Name: string;
    Points: number;
    PrevRank: number;
    Rank: number;
    Rank_Stat: number;
    Rank_Stat_Conquest: any;
    Rank_Stat_Joust: any;
    Rank_Variance: number;
    Round: number;
    Season: number;
    Tier: number;
    Trend: number;
    Wins: number;
    player_id: any;
    ret_msg: any;
  };
  RankedDuelController: {
    Leaves: number;
    Losses: number;
    Name: string;
    Points: number;
    PrevRank: number;
    Rank: number;
    Rank_Stat: number;
    Rank_Stat_Conquest: any;
    Rank_Stat_Joust: any;
    Rank_Variance: number;
    Round: number;
    Season: number;
    Tier: number;
    Trend: number;
    Wins: number;
    player_id: any;
    ret_msg: any;
  };
  RankedJoust: {
    Leaves: number;
    Losses: number;
    Name: string;
    Points: number;
    PrevRank: number;
    Rank: number;
    Rank_Stat: number;
    Rank_Stat_Conquest: any;
    Rank_Stat_Joust: any;
    Rank_Variance: number;
    Round: number;
    Season: number;
    Tier: number;
    Trend: number;
    Wins: number;
    player_id: any;
    ret_msg: any;
  };
  RankedJoustController: {
    Leaves: number;
    Losses: number;
    Name: string;
    Points: number;
    PrevRank: number;
    Rank: number;
    Rank_Stat: number;
    Rank_Stat_Conquest: any;
    Rank_Stat_Joust: any;
    Rank_Variance: number;
    Round: number;
    Season: number;
    Tier: number;
    Trend: number;
    Wins: number;
    player_id: any;
    ret_msg: any;
  };
  Region: string;
  TeamId: number;
  Team_Name: string;
  Tier_Conquest: number;
  Tier_Duel: number;
  Tier_Joust: number;
  Total_Achievements: number;
  Total_Worshippers: number;
  Wins: number;
  hz_gamer_tag: any;
  hz_player_name: string;
  ret_msg: any;
}